OS X: Quiet/fix a lot of deprecated and misc. warnings.

+++ macosx/FileOutlineView.m
@@ -65,7 +65,7 @@
     return [self menu];
 }
 
-- (NSRect) iconRectForRow: (int) row
+- (NSRect) iconRectForRow: (NSInteger) row
 {
     FileNameCell * cell = (FileNameCell *)[self preparedCellAtColumn: [self columnWithIdentifier: @"Name"] row: row];
     NSRect iconRect = [cell imageRectForBounds: [self rectOfRow: row]];
